+++ saleor/graphql/checkout/enums.py
@@ -1,0 +1,7 @@
+import graphene
+
+from ...checkout import error_codes
+
+OrderFromCheckoutCreateErrorCode = graphene.Enum.from_enum(
+    error_codes.OrderFromCheckoutCreateErrorCode
+)
